Close certain files temporarily before forking
This is to work around a new security feature introduced by Google in
the November 16 patches. It would abort the runtime if files are open
that are not whitelisted. To avoid this, we close the files ourselves
and reopen them after Zygote has forked. The mechanism is the same that
Google has introduced, with a few modifications to fit Xposed's needs.

This requires also a change on the native side.

Fixes rovo89/XposedBridge#129.